Add files via upload
I'm learning about cryptography and the simplest one is a Caesar cipher, which we made during a lab. But what happens when we want a shift that's not 3? That's what this shift cipher is for. It takes a string and a shift amount. So if you're given a decrypted sentence, just call this function for a range of (1, 26) and see which one reveals the message.