Create randomizer.py
We are having a scavenger hunt for the CS club and needed to make lists of the clues for everyone. But I didn't want them to be in the same order. So I wrote a Python script that will create random order lists for the desired amount of sheets, place them in a folder, where you can print them all out.